change timing data type from uint8_t to uint16_t

+++ Application/cmd_line_task/command.h
@@ -32,22 +32,22 @@ typedef struct _H_Bridge_task_typedef_
     uint8_t     hv_pos_count;
     uint8_t     hv_neg_count;
 
-    uint8_t     hv_delay_ms;
+    uint16_t    hv_delay_ms;
 
-    uint8_t     hv_pos_on_ms;
-    uint8_t     hv_pos_off_ms;
-    uint8_t     hv_neg_on_ms;
-    uint8_t     hv_neg_off_ms;
+    uint16_t    hv_pos_on_ms;
+    uint16_t    hv_pos_off_ms;
+    uint16_t    hv_neg_on_ms;
+    uint16_t    hv_neg_off_ms;
 
     uint8_t     lv_pos_count;
     uint8_t     lv_neg_count;
 
-    uint8_t     lv_delay_ms;
+    uint16_t    lv_delay_ms;
 
-    uint8_t     lv_pos_on_ms;
-    uint8_t     lv_pos_off_ms;
-    uint8_t     lv_neg_on_ms;
-    uint8_t     lv_neg_off_ms;
+    uint16_t    lv_pos_on_ms;
+    uint16_t    lv_pos_off_ms;
+    uint16_t    lv_neg_on_ms;
+    uint16_t    lv_neg_off_ms;
 } H_Bridge_task_typedef;
 
 /* ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Variables ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ */
